+++ Tests/About.test.ps1
@@ -1,0 +1,7 @@
+﻿
+#requires -Modules Xmlips
+
+task help {
+	. Helps.ps1
+	Test-Helps ..\Module\en-US\Xmlips.dll-Help.ps1
+}